+++ src/constants.js
@@ -1,0 +1,13 @@
+'use strict';
+
+(function () {
+
+  var INDENT = '    ';
+  var LINEBREAK = '\n';
+
+  module.exports = {
+    INDENT: INDENT,
+    LINEBREAK: LINEBREAK
+  };
+
+})();
